refactor(dataset_types): slim bindings + rename LocalFSChromaDB to LocalFileChromaDB
Rewrite the LocalFile and Weaviate bindings against the new
BaseDatasetType contract. Each binding now declares SOURCE_CLS,
VECTOR_STORE_CLS, and split_config(); the default __init__ and
lifecycle delegations come from BaseDatasetType.

LocalFile keeps two binding-specific overrides:
- validate_configuration() injects a UUID collectionName when missing
- ingest() enforces the source's extension allow-list before delegating
  (cross-axis policy belongs on the binding)

Rename LocalFSChromaDBDatasetType to LocalFileChromaDBDatasetType,
move chromadb_local/ to local_file_chromadb/ (directory now describes
the binding pair, not the vector store), and rename chromadb_type.py /
weaviate_type.py to dataset_type.py. Registration paths updated;
public NAME values ("local_file", "remote_weaviate") unchanged so
the frontend dtype continues to resolve.

The chromadb_provisioner.py file moves along with the directory; its
final home in vector_stores/chromadb_local/ is a follow-up.
