+++ backend/syft_space/components/dataset_types/__init__.py
@@ -18,18 +18,18 @@ def register_builtin_types(registry: "DatasetTypeRegistry") -> None:
     # Classes are imported on first use.
     registry.register_lazy_dataset_type(
         "remote_weaviate",
-        "syft_space.components.dataset_types.weaviate_remote.weaviate_type",
+        "syft_space.components.dataset_types.weaviate_remote.dataset_type",
         "RemoteWeaviateDatasetType",
     )
 
     registry.register_lazy_dataset_type(
         "local_file",
-        "syft_space.components.dataset_types.chromadb_local.chromadb_type",
-        "LocalFSChromaDBDatasetType",
+        "syft_space.components.dataset_types.local_file_chromadb.dataset_type",
+        "LocalFileChromaDBDatasetType",
     )
     registry.register_lazy_provisioner(
         "local_file",
-        "syft_space.components.dataset_types.chromadb_local.chromadb_provisioner",
+        "syft_space.components.dataset_types.local_file_chromadb.chromadb_provisioner",
         "LocalChromaDBProvisioner",
     )
