--- backend/syft_space/components/dataset_types/chromadb_local/__init__.py
@@ -1,6 +1,0 @@
-"""ChromaDB local dataset type package."""
-
-from .chromadb_provisioner import LocalChromaDBProvisioner
-from .chromadb_type import LocalFSChromaDBDatasetType
-
-__all__ = ["LocalFSChromaDBDatasetType", "LocalChromaDBProvisioner"]
