+++ backend/syft_space/components/dataset_types/local_file_chromadb/__init__.py
@@ -1,0 +1,6 @@
+"""LocalFile + ChromaDB dataset type binding package."""
+
+from .chromadb_provisioner import LocalChromaDBProvisioner
+from .dataset_type import LocalFileChromaDBDatasetType
+
+__all__ = ["LocalFileChromaDBDatasetType", "LocalChromaDBProvisioner"]
